Add aot_emit_aot_file.h
